# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "43.200.170.45",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/boards/": {
      "post": {
        "tags": [
          "공지글 API"
        ],
        "summary": "공지글 등록 API",
        "description": "-",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "workSpaceName": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/boards/{workSpaceName}": {
      "get": {
        "tags": [
          "공지글 API"
        ],
        "summary": "공지글 전체 조회 API",
        "description": "-",
        "parameters": [
          {
            "name": "workSpaceName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/boards/{boardId}": {
      "put": {
        "tags": [
          "공지글 API"
        ],
        "summary": "공지글 수정 API",
        "description": "-",
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "workSpaceName": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": [
          "공지글 API"
        ],
        "summary": "공지글 삭제 API",
        "description": "-",
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "workSpaceName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/auth/login/kakao/callback": {
      "post": {
        "tags": [
          "카카오 API"
        ],
        "summary": "카카오 콜백 API",
        "description": "-",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/kakao/member": {
      "post": {
        "tags": [
          "카카오 API"
        ],
        "summary": "카카오 정보요청 API",
        "description": "-",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/kakao/parsing": {
      "post": {
        "tags": [
          "카카오 API"
        ],
        "summary": "카카오 파싱 API",
        "description": "-",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user_email": {
                  "example": "any"
                },
                "user_name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/naver": {
      "post": {
        "tags": [
          "네이버 API"
        ],
        "summary": "네이버 콜백 API",
        "description": "-",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/member": {
      "post": {
        "tags": [
          "네이버 API"
        ],
        "summary": "네이버 정보요청 API",
        "description": "-",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/parsing": {
      "post": {
        "tags": [
          "네이버 API"
        ],
        "summary": "네이버 파싱 API",
        "description": "-",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user_email": {
                  "example": "any"
                },
                "user_name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  },
...
}